Correct method to get the number of vertices in a vertex buffer

+++ OpenGL/src/abstraction/VertexBuffer.cpp
@@ -2,7 +2,7 @@
 #include "ErrorHandling.h"
 #include <GL/glew.h>
 
-VertexBuffer::VertexBuffer(const void* data, unsigned int size) : m_count(size)
+VertexBuffer::VertexBuffer(const void* data, size_t size) : m_count(size)
 {
     GLCall(glGenBuffers(1, &m_RendererID));
     GLCall(glBindBuffer(GL_ARRAY_BUFFER, m_RendererID));
@@ -23,7 +23,7 @@ void VertexBuffer::Unbind() const
     GLCall(glBindBuffer(GL_ARRAY_BUFFER, 0));
 }
 
-void VertexBuffer::UpdateData(const void* data, unsigned int size)
+void VertexBuffer::UpdateData(const void* data, size_t size)
 {
     Bind();
     glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
